Add annotation to disable sanitizing secret files

+++ pkg/secrets/pushtofile/provide_conjur_secrets.go
@@ -8,6 +8,7 @@ import (
 	"github.com/cyberark/conjur-opentelemetry-tracer/pkg/trace"
 	"github.com/cyberark/secrets-provider-for-k8s/pkg/log/messages"
 	"github.com/cyberark/secrets-provider-for-k8s/pkg/secrets/clients/conjur"
+	"github.com/cyberark/secrets-provider-for-k8s/pkg/secrets/config"
 	"go.opentelemetry.io/otel"
 )
 
@@ -15,6 +16,7 @@ type fileProvider struct {
 	retrieveSecretsFunc conjur.RetrieveSecretsFunc
 	secretGroups        []*SecretGroup
 	traceContext        context.Context
+	sanitizeEnabled     bool
 }
 
 type fileProviderDepFuncs struct {
@@ -26,11 +28,13 @@ type fileProviderDepFuncs struct {
 // NewProvider creates a new provider for Push-to-File mode.
 func NewProvider(
 	retrieveSecretsFunc conjur.RetrieveSecretsFunc,
-	secretsBasePath string,
-	templatesBasePath string,
-	annotations map[string]string) (*fileProvider, []error) {
+	providerConfig *config.ProviderConfig) (*fileProvider, []error) {
 
-	secretGroups, err := NewSecretGroups(secretsBasePath, templatesBasePath, annotations)
+	secretGroups, err := NewSecretGroups(
+		providerConfig.SecretFileBasePath,
+		providerConfig.TemplateFileBasePath,
+		providerConfig.AnnotationsMap,
+	)
 	if err != nil {
 		return nil, err
 	}
@@ -39,6 +43,7 @@ func NewProvider(
 		retrieveSecretsFunc: retrieveSecretsFunc,
 		secretGroups:        secretGroups,
 		traceContext:        nil,
+		sanitizeEnabled:     providerConfig.SanitizeEnabled,
 	}, nil
 }
 
@@ -47,6 +52,7 @@ func (p fileProvider) Provide() error {
 	return provideWithDeps(
 		p.traceContext,
 		p.secretGroups,
+		p.sanitizeEnabled,
 		fileProviderDepFuncs{
 			retrieveSecretsFunc: p.retrieveSecretsFunc,
 			depOpenWriteCloser:  openFileAsWriteCloser,
@@ -62,6 +68,7 @@ func (p *fileProvider) SetTraceContext(ctx context.Context) {
 func provideWithDeps(
 	traceContext context.Context,
 	groups []*SecretGroup,
+	sanitizeEnabled bool,
 	depFuncs fileProviderDepFuncs,
 ) error {
 	// Use the global TracerProvider
@@ -69,12 +76,13 @@ func provideWithDeps(
 	spanCtx, span := tr.Start(traceContext, "Fetch Conjur Secrets")
 	secretsByGroup, err := FetchSecretsForGroups(depFuncs.retrieveSecretsFunc, groups, spanCtx)
 	if err != nil {
-		// Delete secret files for variables that no longer exist or the user no longer has permissions to
-		// TODO: Should we check the error message to see if it's a 404 or 403?
-		// TODO: Should we only do this when rotation is enabled? If so, how can we determine that?
-		for _, group := range groups {
-			os.Remove(group.FilePath)
-			log.Info(messages.CSPFK019I)
+		if sanitizeEnabled {
+			// Delete secret files for variables that no longer exist or the user no longer has permissions to
+			// TODO: Should we check the error message to see if it's a 404 or 403?
+			for _, group := range groups {
+				os.Remove(group.FilePath)
+				log.Info(messages.CSPFK019I)
+			}
 		}
 
 		span.RecordErrorAndSetStatus(err)
